Add files via upload
Developed a Script using Python, Pandas to scrape every individual details of a shopping app based on the user's search. By gathering all the data and created a data frame which will be helpful for the data visualization, data analysis and data cleaning. Then exported the data frame to a csv file.